[codec] handle failures when decoding utf-8 (#114)
pyspinel fails to parse STREAM_LOG. This is because the format of
STREAM_LOG is "UD" but pyspinel has no idea where the utf-8 part
ends. Therefore it tries to decode the whole data as utf-8 then it
fails when there's an invalid utf-8 character. This commit makes
pyspinel decode utf-8 as far as it can until it meets an invalid
character.